update vivo compat

+++ rom.sh
@@ -36,12 +36,14 @@ if [ -f $OUT/rom/payload.bin ];then
     rm $OUT/rom/payload.bin
     #$MYDIR/tools/deimg.sh $OUT/rom
 else
-    for datsplit in `find out/rom -name  *.new.dat.*|sort --version-sort`;do
+    for datsplit in `find $OUT/rom -name  *.new.dat.*|sort --version-sort`;do
 	path=$(dirname $datsplit)
 	datname=$(basename $datsplit)
 	partname=${datname%%.*}
-	cat $datsplit >> $path/$partname.new.dat
-	rm $datsplit
+        if [ -f $path/$partname.new.dat ];then
+	    cat $datsplit >> $path/$partname.new.dat
+	    rm $datsplit
+        fi
     done
     $MYDIR/tools/dedat.sh $OUT/rom
 fi
